Create a exceptions, UserService, UserSpringSecurity, WebConfig and update the errors in others classes

+++ src/main/java/com/projetominiERP/miniERP/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> { //para busca o id, email...
+public interface UserRepository extends JpaRepository<User, Long> { //para busca o id, email...
 
     @Transactional(readOnly = true)
     User findByEmail(String email); //para buscar o usuário pelo email
